requirements.txt
openai
flask
flask_cors
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv
requests
langchain
langchain_community